telegraf-forwarder/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
idna==3.10
    # via requests
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
urllib3==2.4.0
    # via requests
